refactor: 중복 타입 분리
