+++ react-lunch/src/constants/restaurants.ts
@@ -1,0 +1,11 @@
+export const categoryFilter = [
+  '전체',
+  '한식',
+  '중식',
+  '일식',
+  '양식',
+  '아시안',
+  '기타',
+] as const;
+
+export const alignFilter = ['이름순', '거리순'] as const;
